requirements.txt
# ── Web Framework ──
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# ── LLM & Agent ──
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.3.0
langgraph>=0.2.0
openai>=1.60.0

# ── Data & Validation ──
pydantic>=2.10.0
pyyaml>=6.0.0

# ── Infrastructure ──
python-dotenv>=1.0.1
httpx>=0.28.0
aiosqlite>=0.20.0
redis>=5.2.0

# ── Dev ──
pytest>=8.0.0
pytest-asyncio>=0.24.0
